Refactor: ProblemEntity 수정 timeLimit 타입 변경 : Long -> Integer

+++ src/main/java/ggs/brainvitamin/src/vitamin/entity/ProblemEntity.java
@@ -30,7 +30,7 @@ public class ProblemEntity extends BaseEntity {
     private String trainingName;
 
     @Column(columnDefinition = "INT UNSIGNED NOT NULL")
-    private Long timeLimit;
+    private Integer timeLimit;
 
     @Column(columnDefinition = "VARCHAR(50) NOT NULL")
     private String pathUri;
